Time: 19 ms (8.57%), Space: 2.4 MB (20.00%) - LeetHub

+++ 0645-set-mismatch/0645-set-mismatch.rs
@@ -6,16 +6,16 @@ impl Solution {
         let mut unique_sum = 0;
         let mut total_sum = 0;
         
-        let mut expected_sum = ((nums.len() * (nums.len() + 1))/ 2) as i64;
+        let mut expected_sum = ((nums.len() * (nums.len() + 1))/ 2) as i32;
         
         for n in nums{
             if !set.contains(&n){
-                unique_sum += n as i64;
+                unique_sum += n;
             }
-            total_sum +=  n as i64;
+            total_sum +=  n;
             set.insert(n);
         }
         
-        vec![(total_sum - unique_sum) as i32, (expected_sum - unique_sum) as i32]
+        vec![(total_sum - unique_sum), (expected_sum - unique_sum)]
     }
 }
